Move subscriber and publisher creation to integration tests utils
