+++ integration-testing-app/src/androidTest/java/com/ably/tracking/tests/PublisherAndSubscriberTests.kt
@@ -1,25 +1,13 @@
 package com.ably.tracking.tests
 
-import android.annotation.SuppressLint
-import android.content.Context
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.platform.app.InstrumentationRegistry
-import com.ably.tracking.Accuracy
-import com.ably.tracking.connection.BasicAuthenticationConfiguration
 import com.ably.tracking.LocationUpdate
-import com.ably.tracking.Resolution
-import com.ably.tracking.publisher.DefaultResolutionPolicyFactory
-import com.ably.tracking.publisher.LocationHistoryData
-import com.ably.tracking.publisher.LocationSourceRaw
-import com.ably.tracking.publisher.MapConfiguration
-import com.ably.tracking.publisher.Publisher
-import com.ably.tracking.publisher.RoutingProfile
 import com.ably.tracking.publisher.Trackable
 import com.ably.tracking.subscriber.Subscriber
 import com.ably.tracking.test.common.BooleanExpectation
 import com.ably.tracking.test.common.UnitExpectation
 import com.ably.tracking.test.common.testLogD
-import com.google.gson.Gson
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.launchIn
@@ -31,14 +19,8 @@ import org.junit.Test
 import org.junit.runner.RunWith
 import java.util.UUID
 
-private const val MAPBOX_ACCESS_TOKEN = BuildConfig.MAPBOX_ACCESS_TOKEN
-private const val CLIENT_ID = "IntegrationTestsClient"
-private const val ABLY_API_KEY = BuildConfig.ABLY_API_KEY
-
 @RunWith(AndroidJUnit4::class)
 class PublisherAndSubscriberTests {
-    val gson = Gson()
-
     @Test
     fun createAndStartPublisherAndSubscriberAndWaitUntilDataEnds() {
         // given
@@ -47,7 +29,6 @@ class PublisherAndSubscriberTests {
         val subscriberStoppedExpectation = UnitExpectation("subscriber stopped")
         val context = InstrumentationRegistry.getInstrumentation().targetContext
         val trackableId = UUID.randomUUID().toString()
-        val locationData = getLocationData(context)
         val publishedLocations = mutableListOf<LocationUpdate>()
         val receivedLocations = mutableListOf<LocationUpdate>()
         val trackExpectation = BooleanExpectation("track response")
@@ -68,7 +49,6 @@ class PublisherAndSubscriberTests {
 
         val publisher = createAndStartPublisher(
             context,
-            locationData = locationData,
             onLocationDataEnded = {
                 dataEndedExpectation.fulfill()
             }
@@ -126,36 +106,5 @@ class PublisherAndSubscriberTests {
                 receivedLocation
             )
         }
-    }
-
-    @SuppressLint("MissingPermission")
-    private fun createAndStartPublisher(
-        context: Context,
-        resolution: Resolution = Resolution(Accuracy.BALANCED, 1L, 0.0),
-        locationData: LocationHistoryData,
-        onLocationDataEnded: () -> Unit = {}
-    ) =
-        Publisher.publishers()
-            .androidContext(context)
-            .connection(BasicAuthenticationConfiguration.create(ABLY_API_KEY, CLIENT_ID))
-            .map(MapConfiguration(MAPBOX_ACCESS_TOKEN))
-            .resolutionPolicy(DefaultResolutionPolicyFactory(resolution, context))
-            .profile(RoutingProfile.DRIVING)
-            .locationSource(LocationSourceRaw.create(locationData, onLocationDataEnded))
-            .start()
-
-    private suspend fun createAndStartSubscriber(
-        trackingId: String,
-        resolution: Resolution = Resolution(Accuracy.BALANCED, 1L, 0.0)
-    ) =
-        Subscriber.subscribers()
-            .connection(BasicAuthenticationConfiguration.create(ABLY_API_KEY, CLIENT_ID))
-            .resolution(resolution)
-            .trackingId(trackingId)
-            .start()
-
-    private fun getLocationData(context: Context): LocationHistoryData {
-        val historyString = context.assets.open("location_history_small.txt").use { String(it.readBytes()) }
-        return gson.fromJson(historyString, LocationHistoryData::class.java)
     }
 }
